requirements.txt
liquidpy
rdflib